+++ lib/jenkins2v-ex.c
@@ -41,13 +41,13 @@
 
 #if defined(WEAK_ALIASES) && !defined(__STRICT_ANSI__)
 /* Non-public external symbols for aliasing. */
-void rgph_u32x3_jenkins2v_data32(const void *, size_t, uint32_t, uint32_t *);
-void rgph_u32x3_jenkins2v_data64(const void *, size_t, uint32_t, uint32_t *);
+void rgph_u32x3_jenkins2v_data32(const void *, size_t, uintptr_t, uint32_t *);
+void rgph_u32x3_jenkins2v_data64(const void *, size_t, uintptr_t, uint32_t *);
 #endif
 
 
 void
-rgph_u32x3_jenkins2v_u8(uint8_t value, uint32_t seed, uint32_t *h)
+rgph_u32x3_jenkins2v_u8(uint8_t value, uintptr_t seed, uint32_t *h)
 {
 
 	h[0] = RGPH_JENKINS2V_SEED1;
@@ -60,7 +60,7 @@ rgph_u32x3_jenkins2v_u8(uint8_t value, uint32_t seed, uint32_t *h)
 }
 
 void
-rgph_u32x3_jenkins2v_u16(uint16_t value, uint32_t seed, uint32_t *h)
+rgph_u32x3_jenkins2v_u16(uint16_t value, uintptr_t seed, uint32_t *h)
 {
 
 	h[0] = RGPH_JENKINS2V_SEED1;
@@ -73,7 +73,7 @@ rgph_u32x3_jenkins2v_u16(uint16_t value, uint32_t seed, uint32_t *h)
 }
 
 inline void
-rgph_u32x3_jenkins2v_u32(uint32_t value, uint32_t seed, uint32_t *h)
+rgph_u32x3_jenkins2v_u32(uint32_t value, uintptr_t seed, uint32_t *h)
 {
 
 	h[0] = RGPH_JENKINS2V_SEED1;
@@ -86,7 +86,7 @@ rgph_u32x3_jenkins2v_u32(uint32_t value, uint32_t seed, uint32_t *h)
 }
 
 inline void
-rgph_u32x3_jenkins2v_u64(uint64_t value, uint32_t seed, uint32_t *h)
+rgph_u32x3_jenkins2v_u64(uint64_t value, uintptr_t seed, uint32_t *h)
 {
 
 	h[0] = RGPH_JENKINS2V_SEED1;
@@ -100,14 +100,14 @@ rgph_u32x3_jenkins2v_u64(uint64_t value, uint32_t seed, uint32_t *h)
 }
 
 void
-rgph_u32x3_jenkins2v_f32(float value, uint32_t seed, uint32_t *h)
+rgph_u32x3_jenkins2v_f32(float value, uintptr_t seed, uint32_t *h)
 {
 
 	rgph_u32x3_jenkins2v_u32(rgph_f2u32(value), seed, h);
 }
 
 void
-rgph_u32x3_jenkins2v_f64(double value, uint32_t seed, uint32_t *h)
+rgph_u32x3_jenkins2v_f64(double value, uintptr_t seed, uint32_t *h)
 {
 
 	rgph_u32x3_jenkins2v_u64(rgph_d2u64(value), seed, h);
@@ -115,7 +115,7 @@ rgph_u32x3_jenkins2v_f64(double value, uint32_t seed, uint32_t *h)
 
 inline void
 rgph_u32x3_jenkins2v_data32(const void *data,
-    size_t len, uint32_t seed, uint32_t * restrict h)
+    size_t len, uintptr_t seed, uint32_t * restrict h)
 {
 	const uint32_t * restrict key = data;
 	const uint32_t *end = key + len;
@@ -144,7 +144,7 @@ rgph_u32x3_jenkins2v_data32(const void *data,
 
 inline void
 rgph_u32x3_jenkins2v_data64(const void *data,
-    size_t len, uint32_t seed, uint32_t * restrict h)
+    size_t len, uintptr_t seed, uint32_t * restrict h)
 {
 	const uint64_t * restrict key = data;
 	const uint64_t *end = key + len;
@@ -192,7 +192,7 @@ rgph_u32x3_jenkins2v_data64(const void *data,
 
 void
 rgph_u32x3_jenkins2v_u8a(const uint8_t *key,
-    size_t len, uint32_t seed, uint32_t *h)
+    size_t len, uintptr_t seed, uint32_t *h)
 {
 
 	rgph_u32x3_jenkins2v_data(key, len, seed, h);
@@ -200,7 +200,7 @@ rgph_u32x3_jenkins2v_u8a(const uint8_t *key,
 
 void
 rgph_u32x3_jenkins2v_u16a(const uint16_t *key,
-    size_t len, uint32_t seed, uint32_t *h)
+    size_t len, uintptr_t seed, uint32_t *h)
 {
 	const uint8_t *arg = rgph_unalias(const uint8_t *, key);
 
@@ -208,12 +208,12 @@ rgph_u32x3_jenkins2v_u16a(const uint16_t *key,
 }
 
 #if defined(WEAK_ALIASES) && !defined(__STRICT_ANSI__)
-void rgph_u32x3_jenkins2v_u32a(const uint32_t *, size_t, uint32_t,
+void rgph_u32x3_jenkins2v_u32a(const uint32_t *, size_t, uintptr_t,
     uint32_t *) __attribute__((weak,alias("rgph_u32x3_jenkins2v_data32")));
 #else
 void
 rgph_u32x3_jenkins2v_u32a(const uint32_t *key,
-    size_t len, uint32_t seed, uint32_t *h)
+    size_t len, uintptr_t seed, uint32_t *h)
 {
 
 	rgph_u32x3_jenkins2v_data32(key, len, seed, h);
@@ -221,12 +221,12 @@ rgph_u32x3_jenkins2v_u32a(const uint32_t *key,
 #endif
 
 #if defined(WEAK_ALIASES) && !defined(__STRICT_ANSI__)
-void rgph_u32x3_jenkins2v_u64a(const uint64_t *, size_t, uint32_t,
+void rgph_u32x3_jenkins2v_u64a(const uint64_t *, size_t, uintptr_t,
     uint32_t *) __attribute__((weak,alias("rgph_u32x3_jenkins2v_data64")));
 #else
 void
 rgph_u32x3_jenkins2v_u64a(const uint64_t *key,
-    size_t len, uint32_t seed, uint32_t *h)
+    size_t len, uintptr_t seed, uint32_t *h)
 {
 
 	rgph_u32x3_jenkins2v_data64(key, len, seed, h);
@@ -234,12 +234,12 @@ rgph_u32x3_jenkins2v_u64a(const uint64_t *key,
 #endif
 
 #if defined(WEAK_ALIASES) && !defined(__STRICT_ANSI__)
-void rgph_u32x3_jenkins2v_f32a(const float *, size_t, uint32_t,
+void rgph_u32x3_jenkins2v_f32a(const float *, size_t, uintptr_t,
     uint32_t *) __attribute__((weak,alias("rgph_u32x3_jenkins2v_data32")));
 #else
 void
 rgph_u32x3_jenkins2v_f32a(const float *key,
-    size_t len, uint32_t seed, uint32_t *h)
+    size_t len, uintptr_t seed, uint32_t *h)
 {
 
 	rgph_u32x3_jenkins2v_data32(key, len, seed, h);
@@ -247,12 +247,12 @@ rgph_u32x3_jenkins2v_f32a(const float *key,
 #endif
 
 #if defined(WEAK_ALIASES) && !defined(__STRICT_ANSI__)
-void rgph_u32x3_jenkins2v_f64a(const double *, size_t, uint32_t,
+void rgph_u32x3_jenkins2v_f64a(const double *, size_t, uintptr_t,
     uint32_t *) __attribute__((weak,alias("rgph_u32x3_jenkins2v_data64")));
 #else
 void
 rgph_u32x3_jenkins2v_f64a(const double *key,
-    size_t len, uint32_t seed, uint32_t *h)
+    size_t len, uintptr_t seed, uint32_t *h)
 {
 
 	rgph_u32x3_jenkins2v_data64(key, len, seed, h);
