+++ lib/murmur32v-ex.c
@@ -38,13 +38,13 @@
 
 #if defined(WEAK_ALIASES) && !defined(__STRICT_ANSI__)
 /* Non-public external symbols for aliasing. */
-void rgph_u32x4_murmur32v_data32(const void *, size_t, uint32_t, uint32_t *);
-void rgph_u32x4_murmur32v_data64(const void *, size_t, uint32_t, uint32_t *);
+void rgph_u32x4_murmur32v_data32(const void *, size_t, uintptr_t, uint32_t *);
+void rgph_u32x4_murmur32v_data64(const void *, size_t, uintptr_t, uint32_t *);
 #endif
 
 
 void
-rgph_u32x4_murmur32v_u8(uint8_t value, uint32_t seed, uint32_t *h)
+rgph_u32x4_murmur32v_u8(uint8_t value, uintptr_t seed, uint32_t *h)
 {
 
 	h[0] = h[1] = h[2] = h[3] = seed;
@@ -54,7 +54,7 @@ rgph_u32x4_murmur32v_u8(uint8_t value, uint32_t seed, uint32_t *h)
 }
 
 void
-rgph_u32x4_murmur32v_u16(uint16_t value, uint32_t seed, uint32_t *h)
+rgph_u32x4_murmur32v_u16(uint16_t value, uintptr_t seed, uint32_t *h)
 {
 
 	h[0] = h[1] = h[2] = h[3] = seed;
@@ -64,7 +64,7 @@ rgph_u32x4_murmur32v_u16(uint16_t value, uint32_t seed, uint32_t *h)
 }
 
 inline void
-rgph_u32x4_murmur32v_u32(uint32_t value, uint32_t seed, uint32_t *h)
+rgph_u32x4_murmur32v_u32(uint32_t value, uintptr_t seed, uint32_t *h)
 {
 
 	h[0] = h[1] = h[2] = h[3] = seed;
@@ -74,7 +74,7 @@ rgph_u32x4_murmur32v_u32(uint32_t value, uint32_t seed, uint32_t *h)
 }
 
 inline void
-rgph_u32x4_murmur32v_u64(uint64_t value, uint32_t seed, uint32_t *h)
+rgph_u32x4_murmur32v_u64(uint64_t value, uintptr_t seed, uint32_t *h)
 {
 
 	h[0] = h[1] = h[2] = h[3] = seed;
@@ -85,14 +85,14 @@ rgph_u32x4_murmur32v_u64(uint64_t value, uint32_t seed, uint32_t *h)
 }
 
 void
-rgph_u32x4_murmur32v_f32(float value, uint32_t seed, uint32_t *h)
+rgph_u32x4_murmur32v_f32(float value, uintptr_t seed, uint32_t *h)
 {
 
 	rgph_u32x4_murmur32v_u32(rgph_f2u32(value), seed, h);
 }
 
 void
-rgph_u32x4_murmur32v_f64(double value, uint32_t seed, uint32_t *h)
+rgph_u32x4_murmur32v_f64(double value, uintptr_t seed, uint32_t *h)
 {
 
 	rgph_u32x4_murmur32v_u64(rgph_d2u64(value), seed, h);
@@ -100,7 +100,7 @@ rgph_u32x4_murmur32v_f64(double value, uint32_t seed, uint32_t *h)
 
 inline void
 rgph_u32x4_murmur32v_data32(const void *data,
-    size_t len, uint32_t seed, uint32_t * restrict h)
+    size_t len, uintptr_t seed, uint32_t * restrict h)
 {
 	const uint32_t * restrict key = data;
 	const uint32_t *end = key + len;
@@ -133,7 +133,7 @@ rgph_u32x4_murmur32v_data32(const void *data,
 
 inline void
 rgph_u32x4_murmur32v_data64(const void *data,
-    size_t len, uint32_t seed, uint32_t * restrict h)
+    size_t len, uintptr_t seed, uint32_t * restrict h)
 {
 	const uint64_t * restrict key = data;
 	const uint64_t *end = key + len;
@@ -166,7 +166,7 @@ rgph_u32x4_murmur32v_data64(const void *data,
 
 void
 rgph_u32x4_murmur32v_u8a(const uint8_t *key,
-    size_t len, uint32_t seed, uint32_t *h)
+    size_t len, uintptr_t seed, uint32_t *h)
 {
 
 	rgph_u32x4_murmur32v_data(key, len, seed, h);
@@ -174,7 +174,7 @@ rgph_u32x4_murmur32v_u8a(const uint8_t *key,
 
 void
 rgph_u32x4_murmur32v_u16a(const uint16_t *key,
-    size_t len, uint32_t seed, uint32_t *h)
+    size_t len, uintptr_t seed, uint32_t *h)
 {
 	const uint8_t *arg = rgph_unalias(const uint8_t *, key);
 
@@ -182,12 +182,12 @@ rgph_u32x4_murmur32v_u16a(const uint16_t *key,
 }
 
 #if defined(WEAK_ALIASES) && !defined(__STRICT_ANSI__)
-void rgph_u32x4_murmur32v_u32a(const uint32_t *, size_t, uint32_t,
+void rgph_u32x4_murmur32v_u32a(const uint32_t *, size_t, uintptr_t,
     uint32_t *) __attribute__((weak,alias("rgph_u32x4_murmur32v_data32")));
 #else
 void
 rgph_u32x4_murmur32v_u32a(const uint32_t *key,
-    size_t len, uint32_t seed, uint32_t *h)
+    size_t len, uintptr_t seed, uint32_t *h)
 {
 
 	rgph_u32x4_murmur32v_data32(key, len, seed, h);
@@ -195,12 +195,12 @@ rgph_u32x4_murmur32v_u32a(const uint32_t *key,
 #endif
 
 #if defined(WEAK_ALIASES) && !defined(__STRICT_ANSI__)
-void rgph_u32x4_murmur32v_u64a(const uint64_t *, size_t, uint32_t,
+void rgph_u32x4_murmur32v_u64a(const uint64_t *, size_t, uintptr_t,
     uint32_t *) __attribute__((weak,alias("rgph_u32x4_murmur32v_data64")));
 #else
 void
 rgph_u32x4_murmur32v_u64a(const uint64_t *key,
-    size_t len, uint32_t seed, uint32_t *h)
+    size_t len, uintptr_t seed, uint32_t *h)
 {
 
 	rgph_u32x4_murmur32v_data64(key, len, seed, h);
@@ -208,12 +208,12 @@ rgph_u32x4_murmur32v_u64a(const uint64_t *key,
 #endif
 
 #if defined(WEAK_ALIASES) && !defined(__STRICT_ANSI__)
-void rgph_u32x4_murmur32v_f32a(const float *, size_t, uint32_t,
+void rgph_u32x4_murmur32v_f32a(const float *, size_t, uintptr_t,
     uint32_t *) __attribute__((weak,alias("rgph_u32x4_murmur32v_data32")));
 #else
 void
 rgph_u32x4_murmur32v_f32a(const float *key,
-    size_t len, uint32_t seed, uint32_t *h)
+    size_t len, uintptr_t seed, uint32_t *h)
 {
 
 	return rgph_u32x4_murmur32v_data32(key, len, seed, h);
@@ -221,12 +221,12 @@ rgph_u32x4_murmur32v_f32a(const float *key,
 #endif
 
 #if defined(WEAK_ALIASES) && !defined(__STRICT_ANSI__)
-void rgph_u32x4_murmur32v_f64a(const double *, size_t, uint32_t,
+void rgph_u32x4_murmur32v_f64a(const double *, size_t, uintptr_t,
     uint32_t *) __attribute__((weak,alias("rgph_u32x4_murmur32v_data64")));
 #else
 void
 rgph_u32x4_murmur32v_f64a(const double *key,
-    size_t len, uint32_t seed, uint32_t *h)
+    size_t len, uintptr_t seed, uint32_t *h)
 {
 
 	return rgph_u32x4_murmur32v_data64(key, len, seed, h);
