+++ lib/xxh32s-ex.c
@@ -39,13 +39,13 @@
 
 #if defined(WEAK_ALIASES) && !defined(__STRICT_ANSI__)
 /* Non-public external symbols for aliasing. */
-uint32_t rgph_u32_xxh32s_data32(const void *, size_t, uint32_t);
-uint32_t rgph_u32_xxh32s_data64(const void *, size_t, uint32_t);
+uint32_t rgph_u32_xxh32s_data32(const void *, size_t, uintptr_t);
+uint32_t rgph_u32_xxh32s_data64(const void *, size_t, uintptr_t);
 #endif
 
 
 uint32_t
-rgph_u32_xxh32s_u8(uint8_t value, uint32_t seed)
+rgph_u32_xxh32s_u8(uint8_t value, uintptr_t seed)
 {
 	uint32_t h[4];
 
@@ -59,7 +59,7 @@ rgph_u32_xxh32s_u8(uint8_t value, uint32_t seed)
 }
 
 uint32_t
-rgph_u32_xxh32s_u16(uint16_t value, uint32_t seed)
+rgph_u32_xxh32s_u16(uint16_t value, uintptr_t seed)
 {
 	uint32_t h[4];
 
@@ -74,7 +74,7 @@ rgph_u32_xxh32s_u16(uint16_t value, uint32_t seed)
 }
 
 inline uint32_t
-rgph_u32_xxh32s_u32(uint32_t value, uint32_t seed)
+rgph_u32_xxh32s_u32(uint32_t value, uintptr_t seed)
 {
 	uint32_t h[4];
 
@@ -88,7 +88,7 @@ rgph_u32_xxh32s_u32(uint32_t value, uint32_t seed)
 }
 
 inline uint32_t
-rgph_u32_xxh32s_u64(uint64_t value, uint32_t seed)
+rgph_u32_xxh32s_u64(uint64_t value, uintptr_t seed)
 {
 	uint32_t h[4];
 
@@ -103,21 +103,21 @@ rgph_u32_xxh32s_u64(uint64_t value, uint32_t seed)
 }
 
 uint32_t
-rgph_u32_xxh32s_f32(float value, uint32_t seed)
+rgph_u32_xxh32s_f32(float value, uintptr_t seed)
 {
 
 	return rgph_u32_xxh32s_u32(rgph_f2u32(value), seed);
 }
 
 uint32_t
-rgph_u32_xxh32s_f64(double value, uint32_t seed)
+rgph_u32_xxh32s_f64(double value, uintptr_t seed)
 {
 
 	return rgph_u32_xxh32s_u64(rgph_d2u64(value), seed);
 }
 
 inline uint32_t
-rgph_u32_xxh32s_data32(const void *data, size_t len, uint32_t seed)
+rgph_u32_xxh32s_data32(const void *data, size_t len, uintptr_t seed)
 {
 	const uint32_t *key = data;
 	const uint32_t *end = key + len;
@@ -143,7 +143,7 @@ rgph_u32_xxh32s_data32(const void *data, size_t len, uint32_t seed)
 }
 
 inline uint32_t
-rgph_u32_xxh32s_data64(const void *data, size_t len, uint32_t seed)
+rgph_u32_xxh32s_data64(const void *data, size_t len, uintptr_t seed)
 {
 	const uint64_t *key = data;
 	const uint64_t *end = key + len;
@@ -171,14 +171,14 @@ rgph_u32_xxh32s_data64(const void *data, size_t len, uint32_t seed)
 }
 
 uint32_t
-rgph_u32_xxh32s_u8a(const uint8_t *key, size_t len, uint32_t seed)
+rgph_u32_xxh32s_u8a(const uint8_t *key, size_t len, uintptr_t seed)
 {
 
 	return rgph_u32_xxh32s_data(key, len, seed);
 }
 
 uint32_t
-rgph_u32_xxh32s_u16a(const uint16_t *key, size_t len, uint32_t seed)
+rgph_u32_xxh32s_u16a(const uint16_t *key, size_t len, uintptr_t seed)
 {
 	const uint8_t *arg = rgph_unalias(const uint8_t *, key);
 
@@ -187,10 +187,10 @@ rgph_u32_xxh32s_u16a(const uint16_t *key, size_t len, uint32_t seed)
 
 #if defined(WEAK_ALIASES) && !defined(__STRICT_ANSI__)
 uint32_t rgph_u32_xxh32s_u32a(const uint32_t *, size_t,
-    uint32_t) __attribute__((weak,alias("rgph_u32_xxh32s_data32")));
+    uintptr_t) __attribute__((weak,alias("rgph_u32_xxh32s_data32")));
 #else
 uint32_t
-rgph_u32_xxh32s_u32a(const uint32_t *key, size_t len, uint32_t seed)
+rgph_u32_xxh32s_u32a(const uint32_t *key, size_t len, uintptr_t seed)
 {
 
 	return rgph_u32_xxh32s_data32(key, len, seed);
@@ -199,10 +199,10 @@ rgph_u32_xxh32s_u32a(const uint32_t *key, size_t len, uint32_t seed)
 
 #if defined(WEAK_ALIASES) && !defined(__STRICT_ANSI__)
 uint32_t rgph_u32_xxh32s_u64a(const uint64_t *, size_t,
-    uint32_t) __attribute__((weak,alias("rgph_u32_xxh32s_data64")));
+    uintptr_t) __attribute__((weak,alias("rgph_u32_xxh32s_data64")));
 #else
 uint32_t
-rgph_u32_xxh32s_u64a(const uint64_t *key, size_t len, uint32_t seed)
+rgph_u32_xxh32s_u64a(const uint64_t *key, size_t len, uintptr_t seed)
 {
 
 	return rgph_u32_xxh32s_data64(key, len, seed);
@@ -211,10 +211,10 @@ rgph_u32_xxh32s_u64a(const uint64_t *key, size_t len, uint32_t seed)
 
 #if defined(WEAK_ALIASES) && !defined(__STRICT_ANSI__)
 uint32_t rgph_u32_xxh32s_f32a(const float *, size_t,
-    uint32_t) __attribute__((weak,alias("rgph_u32_xxh32s_data32")));
+    uintptr_t) __attribute__((weak,alias("rgph_u32_xxh32s_data32")));
 #else
 uint32_t
-rgph_u32_xxh32s_f32a(const float *key, size_t len, uint32_t seed)
+rgph_u32_xxh32s_f32a(const float *key, size_t len, uintptr_t seed)
 {
 
 	return rgph_u32_xxh32s_data32(key, len, seed);
@@ -223,10 +223,10 @@ rgph_u32_xxh32s_f32a(const float *key, size_t len, uint32_t seed)
 
 #if defined(WEAK_ALIASES) && !defined(__STRICT_ANSI__)
 uint32_t rgph_u32_xxh32s_f64a(const double *, size_t,
-    uint32_t) __attribute__((weak,alias("rgph_u32_xxh32s_data64")));
+    uintptr_t) __attribute__((weak,alias("rgph_u32_xxh32s_data64")));
 #else
 uint32_t
-rgph_u32_xxh32s_f64a(const double *key, size_t len, uint32_t seed)
+rgph_u32_xxh32s_f64a(const double *key, size_t len, uintptr_t seed)
 {
 
 	return rgph_u32_xxh32s_data64(key, len, seed);
