+++ lib/xxh64s.c
@@ -39,7 +39,7 @@
 
 
 uint64_t
-rgph_u64_xxh64s_data(const void *data, size_t len, uint32_t seed)
+rgph_u64_xxh64s_data(const void *data, size_t len, uintptr_t seed)
 {
 	const uint8_t *key = data;
 	const uint8_t *end = key + len;
